refactor and file reorganization
