--- src/Core/Content/Product/SalesChannel/AbstractExampleRoute.php
@@ -1,13 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Torq\Shopware\DynamicAccessEvolved\Core\Content\Product\SalesChannel;
-
-use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
-use Shopware\Core\System\SalesChannel\SalesChannelContext;
-
-abstract class AbstractExampleRoute
-{
-    abstract public function getDecorated(): AbstractExampleRoute;
-
-    abstract public function load(Criteria $criteria, SalesChannelContext $context): ExampleRouteResponse;
-}
