--- src/Core/Content/Product/SalesChannel/ExampleRouteResponse.php
@@ -1,21 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Torq\Shopware\DynamicAccessEvolved\Core\Content\Product\SalesChannel;
-
-use Shopware\Core\Content\Product\ProductCollection;
-use Shopware\Core\Framework\DataAbstractionLayer\Search\EntitySearchResult;
-use Shopware\Core\System\SalesChannel\StoreApiResponse;
-
-/**
- * @property EntitySearchResult $object
- */
-class ExampleRouteResponse extends StoreApiResponse
-{
-    public function getExamples(): ProductCollection
-    {
-        /** @var ProductCollection $collection */
-        $collection = $this->object->getEntities();
-
-        return $collection;
-    }
-}
